requirements.txt
llama-cpp-python
langchain
#chromadb
pinecone-client
#pyopenssl
#pgvector
#psycopg2
